Added pathcomplete markers
Fixed bug with agent planning. This was caused by not resetting the STAStar graph when planning the next agent, instead the graph was just growing larger.
When an agent reaches its destination, it will get a new destination and attempt to plan to it. This is done by tracking the timesteps of each agent.
Currently, if an agent fails to find a path, it will wait in it's current location for a timestep and attempt to plan to a new destination in the next timestep.