Fixed CBS bug so it is now functional
CBS is now functional (apart from STAStar wait action).
The bug was due to newNode.constraints = workingNode.constraints; this was resulting in the ConflictTreeNodes all sharing the same constraints as it is by reference, fix was to add constraints manually.
Changed STAStar to return a path including the agent starting node so edge collisions at time 0 can be detected.